// apps/client/src/pages/confeti/types/festival.ts
export interface FestivalData {
  festivalId: number;
  posterUrl: string;
  posterBgUrl: string;
  title: string;
  subtitle: string;
  startAt: string;
  endAt: string;
  area: string;
  reserveAt: string;
  reservationUrl: string;
  time: string;
  ageRating: string;
  reservationOffice: string;
  price: string;
  infoImgUrl: string;
  isFavorite: boolean;
}

export interface FestivalDate {
  festivalDateId: number;
  festivalAt: string;
  isOpen: boolean;
  artists: {
    artistId: string;
    name: string;
    profileUrl: string;
  }[];
}